+++ virt/virt-manager/virt-install/uefi.sh
@@ -1,7 +1,9 @@
 #!/bin/bash
 
+uefi_args=( --boot uefi )
+
 # https://fedoraproject.org/wiki/Using_UEFI_with_QEMU
 sudo virt-install --name f20-uefi \
 	--ram 2048 --disk size=20 \
-	--boot uefi \
+	${uefi_args[@]} \
 	--location https://dl.fedoraproject.org/pub/fedora/linux/releases/22/Workstation/x86_64/os/